pyproject.toml
[tool.poetry]
name = "modelos"
version = "0.1.0"
description = "An operating system for machine learning"
authors = ["Patrick Barker <patrickbarkerco@gmail.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.8"
removestar = "^1.3.1"
unimport = "^0.14.1"
isort = "^5.12.0"
black = "^23.1.0"
docker = "^6.0.1"
kubernetes = "^25.3.0"
starlette = "^0.24.0"
uvicorn = "^0.20.0"
cloudpickle = "^2.2.1"
dataclasses-jsonschema = "^2.16.0"
docker-image-py = "^0.1.12"
GitPython = "^3.1.30"
opencontainers = "^0.0.14"
deepdiff = "^6.2.3"
semver = "^2.13.0"
xdg = "^5.1.1"
ocifacts = "^0.0.9"

[tool.poetry.dev-dependencies]
mypy = "^1.0.0"
pylint = "^2.16.1"
types-PyYAML = "^6.0.12"
types-six = "^1.16.21"
ipykernel = "^6.21.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
